Torrentday: Switched back to the new rest api.
